Added permisions for users to admin site to User model
