+++ ProjectShop/UserApp/forms.py
@@ -1,0 +1,47 @@
+from django import forms
+from django.contrib.auth.forms import ReadOnlyPasswordHashField
+from django.core.exceptions import ValidationError
+
+from UserApp.models import User
+
+
+class UserCreationForm(forms.ModelForm):
+    """
+        A form for creating new users. Includes all the required
+        fields, plus a repeated password.
+    """
+    password1 = forms.CharField(label='Password', widget=forms.PasswordInput)
+    password2 = forms.CharField(label='Password confirmation', widget=forms.PasswordInput)
+
+    class Meta:
+        model = User
+        fields = ('phone_number', 'first_name', 'last_name', 'middle_name', 'birth_date', 'email', 'role')
+
+    def clean_password2(self):
+        # Check that the two password entries match
+        password1 = self.cleaned_data.get("password1")
+        password2 = self.cleaned_data.get("password2")
+        if password1 and password2 and password1 != password2:
+            raise ValidationError("Passwords don't match")
+        return password2
+
+    def save(self, commit=True):
+        # Save the provided password in hashed format
+        user = super().save(commit=False)
+        user.set_password(self.cleaned_data["password1"])
+        if commit:
+            user.save()
+
+        return user
+
+
+class UserChangeForm(forms.ModelForm):
+    """
+        A form for updating users. Includes all the fields on
+        the user that are written in Meta.fields
+    """
+    password = ReadOnlyPasswordHashField()
+
+    class Meta:
+        model = User
+        fields = ('first_name', 'last_name', 'middle_name', 'birth_date', 'phone_number', 'email', 'role')
